Fix bad selector in sponsor button handler.
This also seems to be a subtle difference between kramdown and Maruku.
Maruku appears to generate header ids using underscores, and kramdown
hyphens. This selector depends on the automatically generated id, which
unfortunately makes it susceptible to breakage.

However, functionality exists in kramdown (and has existed in Maruku) to
explicitly specify the id of a header. This might be worth implementing
in the interest of robustifying this code.

http://kramdown.gettalong.org/syntax.html#specifying-a-header-id