Split out publications and added newsletter redirect (#202)
Publications were mixed in with the other team documents and tutorials. After discussing with marketing, we decided to split publications out and make it easier to find. Also, it was annoying to navigate through the newsletters in the blog, so we allowed blog posts to link directly to an external resource such as newsletters.